airflow/requirements.txt
kafka-python>=2.0.2
requests>=2.25.0
python-dotenv>=0.19.0 